Working on JSON files.

+++ src/main/java/org/grible/servlets/app/delete/DeleteTable.java
@@ -19,6 +19,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.grible.security.Security;
 import org.grible.servlets.ServletHelper;
 
 /**
@@ -43,6 +44,9 @@ public class DeleteTable extends HttpServlet {
 		response.setContentType("text/html");
 		PrintWriter out = response.getWriter();
 		try {
+			if (Security.anyServletEntryCheckFailed(request, response)) {
+				return;
+			}
 			int tableId = Integer.parseInt(request.getParameter("id"));
 			out.print(ServletHelper.deleteTable(tableId));
 		} catch (Exception e) {
